Plan Details Management

+++ app/Http/Controllers/Admin/PlanController.php
@@ -51,9 +51,19 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+
+        $plan = $this->repository->with('details')
+            ->where('url', $url)
+            ->first();
         if (!$plan)
             redirect()->back();
+
+
+        if ($plan->details()->count() > 0) {
+            return redirect()->back()
+                ->with('error', 'Existem detalhes vinculados a esse plano, portanto não pode deletar.');
+        }
+
         $plan->delete();
         return redirect()->route('plans.index');
     }
